requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
pytest-flask==1.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
PyMySQL==1.1.0
mysqlclient==2.2.0
cryptography==41.0.4
uuid==1.30
jsonschema==4.19.1